Moved intersection logic to state drawings themselves
Two new methods have been added: getIntersection and getIntersectionAt.
getIntersection retrieves the intersection point on the border of the
state drawing when shooting a vector from the center of the state to
that point. getIntersectionAt does the same thing, but instead requires
an angle (in radians) to find the intersection point.