Added new Snappable interface. StateDrawings implements this interface. It works
The new Snappable interface is implemented by GraphDrawing and StateDrawing.
When the drawer starts drawing the snap method is called if the drawing is
snappable. The snap method for GraphDrawing snaps each state drawing into
place. The snapToGrid method has been removed, though its implementation
is equal to the snap method for the state drawings. This has the consequence
that every drawing object is responsible for snapping itself into place,
providing us with more flexibility when it comes to defining what is the
"right" way to snap.

It is recommendable to extend the DrawerOptions object to include options
regarding whether to snap drawings, and whether to draw a grid, and if so,
with which parameters.