Moved to building with webpack, using NPM.
This was a gargantuan commit. Virtually all files have been moved
or had to be edited in some way to work with the new architecture.
One annoying part of the new structure is that every single class
needs to be imported individually. Will have to look into TypeScript
namespaces. Hopefully this can be done easier that way.