Added html-webpack-plugin. Dist folder no longer necessary
The html-webpack-plugin generates a new html file in the dist folder
when npm run build is run. The compiled css file is also put there.
Everything in the dist folder is therefore automatically generated and
we do not need it in the repo anymore. The .gitignore file has been
modified accordingly. Furthermore, the template compilation for vue
components now only take place for html files in the src/components/
directory.