Added session service. Changed the upload component accordingly.
To start the modeller the uid, sid, and pid need to be known. The sid
can only be determined when the uid and pid are known. Because the pid
is gained by requesting one from the server from the upload component,
we need to know when the pid is actually received. Once this is done
we can request the sid. To do this the actions now return a promise, so
that we can determine from the components themselves whether the request
is done, and whether it succeeded or failed.