Renamed the User store to the session stores as this is a more appropriate name
The session store stores all information that is needed for the session. Right now
this is the user id, the Petri net id, and the session id. The graph, Petri net and
feedback might be added to this. Although a seperate store module for the feedback
may be a better solution.